assignment_2/requirements.txt
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
flask_cors==3.0.10
gensim==4.3.2
gunicorn==21.2.0
h11==0.14.0
idna==3.4
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.9
nltk==3.8.1
numpy==1.26.2
openpyxl==3.1.2
outcome==1.3.0.post0
package_name==0.1
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.2
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.16.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.0.1
wsproto==1.2.0